mypy fix
