+++ config/language.validate.py
@@ -1,29 +1,19 @@
-import yaml
+from typing import Dict, List
+from module.utils import load_yaml, error
+from module.validation import validate_item_structure
 
 FILE = "language.yaml"
+LANGUAGES = ["en", "fa"]
 
-file = open(FILE, "r", encoding="utf-8")
-docs = yaml.safe_load_all(file)
-
-
-def error(msg: str) -> None:
-    print("Error: " + msg)
-    exit(1)
-
+docs: List[Dict] = load_yaml(FILE)
 
 if __name__ == "__main__":
     for doc in docs:
-        for item in doc["items"]:
-            if "id" not in item:
+        for item in doc.get("items", []):
+            try:
+                validate_item_structure(item, LANGUAGES)
+            except ValueError as e:
                 print(item)
-                error("id is required and missed in an item")
-
-            if "name" not in item:
-                print(item)
-                error("name is required and missed in an item")
-
-            if "local_name" not in item:
-                print(item)
-                error("local_name is required and missed in an item")
+                error(str(e))
 
     print(FILE + ": Validation is successful")
