+++ config/layout/attribute/type.validate.py
@@ -5,13 +5,15 @@ sys.path.append(
     os.path.abspath(os.path.join(os.path.dirname(__file__), "../..", "module"))
 )
 
-from utils import error, load_yaml, validate_item_structure
+from utils import error, load_yaml
+from validation import validate_item_structure
+
 
 FILE = "type.yaml"
 LANGUAGES = ["en", "fa"]
 
 
-def main():
+def main() -> None:  # Add the return type annotation here
     try:
         docs = load_yaml(FILE)
 
@@ -30,4 +32,4 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    main()  # Call the main function
